requirements.txt
google-genai
python-dotenv
Pillow
gradio_client   
pyautogui
pynput